pyproject.toml
[project]
name = "df-translation-stats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langcodes>=3.5.0",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "scour>=0.38.2",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "C408",
    "D",
]

[tool.poe.tasks]
run.script = "df_translation_stats.main:main"
